[backend] Add Zulip Backend
This commit adds support to perceval to fetch messages
from a Zulip server stream. The user can create a bot
for the email and api_token, required for authentication.

The backend can be used as
$ perceval zulip '<URL>' '<STREAM>' -e '<EMAIL>' -t '<API_KEY>'

The tests have been added accordingly and the readme
is updated with the instructions to install and use the
perceval backend.

The dependencies are added using poetry and the flake8
configurations are also added.

Signed-off-by: Venu Vardhan Reddy Tekula <venu@chaoss.community>