[poetry] Update the dependencies
Poetry doesn't publish packages with git source URL
as dependencies. The grimoirelab related packages
need to have git URLs causing error with the workflows.

This commit implements a workaround to make the backend
work and allows poetry to publish the package without
any trouble.

It adds the grimoirelab dependencies as * in
[tool.poetry.dependencies] and git URL in the
[tool.poetry.dev-dependencies] section.

Signed-off-by: Venu Vardhan Reddy Tekula <venu@chaoss.community>